Upgrade to OpenJFX 12
