+++ .gitignore
@@ -1,12 +1,5 @@
-/VFXWindows-Tutorial-1/nbproject/private/
-/VFXWindows-Tutorial-1/build/
-/VFXWindows-Tutorial-1/dist/
-/VFXWindows-Tutorial-2/nbproject/private/
-/VFXWindows-Tutorial-2/build/
-/VFXWindows-Tutorial-2/dist/
-/VFXWindows-Tutorial-3/nbproject/private/
-/VFXWindows-Tutorial-3/build/
-/VFXWindows-Tutorial-3/dist/
-/VFXWindows-Tutorial-3_1/nbproject/private/
-/VFXWindows-Tutorial-3_1/dist/
-/VFXWindows-Tutorial-3_1/build/
+**/nbproject/private/
+**/build/
+**/dist/
+**/.idea
+*.iml
