+++ README.md
@@ -3,5 +3,22 @@ VFXWindows-Samples
 
 Tutorial applications for JFXtras/VFXWindows library.
 
+## Tutorial 1
+Shows a single window inside the main window.
+
+Needs the following JVM options `--add-exports=jfxtras.window/jfxtras.internal.scene.control.skin.window=javafx.controls`
+
+## Tutorial 2
+Shows a single window inside the main window,
+decorated with buttons.
+
+Needs the following JVM options `--add-exports=jfxtras.window/jfxtras.internal.scene.control.skin.window=javafx.controls`
+
+## Tutorial 3
+Currently not functioning due to resource loading issues.
+
+## Tutorial 3.1
+Currently not functioning due to resource loading issues.
+
 > **NOTE:** VFXWindows has been added to [JFXtras](https://github.com/JFXtras/jfxtras-labs).
 All samples have been updated to work with JFXtras. Please don't use the old VFXWindows library anymore.
