--- VFXWindows-Tutorial-1/manifest.mf
@@ -1,3 +1,0 @@
-Manifest-Version: 1.0
-X-COMMENT: Main-Class will be added automatically by build
-
