--- VFXWindows-Tutorial-1/nbproject/genfiles.properties
@@ -1,8 +1,0 @@
-build.xml.data.CRC32=bc93ac14
-build.xml.script.CRC32=56ee5423
-build.xml.stylesheet.CRC32=28e38971@1.53.1.46
-# This file is used by a NetBeans-based IDE to track changes in generated files such as build-impl.xml.
-# Do not edit this file. You may delete it but then the IDE will never regenerate such files for you.
-nbproject/build-impl.xml.data.CRC32=1d3b58fc
-nbproject/build-impl.xml.script.CRC32=9d356d5a
-nbproject/build-impl.xml.stylesheet.CRC32=6ddba6b6@1.53.1.46
