+++ VFXWindows-Tutorial-2/src/main/java/module-info.java
@@ -1,0 +1,7 @@
+module vfx_windows_tutorial_2 {
+    requires javafx.controls;
+    requires jfxtras.window;
+
+    opens eu.mihosoft.vrl.fxwindows.tutorial to javafx.fxml;
+    exports eu.mihosoft.vrl.fxwindows.tutorial;
+}
