+++ VFXWindows-Tutorial-3_1/src/main/java/module-info.java
@@ -1,0 +1,8 @@
+module vfx_windows_tutorial_3_1 {
+    requires javafx.controls;
+    requires javafx.web;
+    requires jfxtras.window;
+
+    opens eu.mihosoft.vrl.fxwindows.tutorial to javafx.fxml;
+    exports eu.mihosoft.vrl.fxwindows.tutorial;
+}
